// src/types/SanggahanTypes.tsx
export type ResponseSanggahanMhsPaginationType = {
    PendaftaranId: string
    NamaProgramStudi: string
    KodePendaftar: string
    NoUjian: string
    Periode: string
    MataKuliah: number
    Status: string
    SanggahanAssesmenId: string
    SanggahanProsesBanding: boolean
    SanggahanDiskusiBanding: boolean
}

export type ResponseSanggahanAsesorPaginationType = {
    PendaftaranId: string
    Nama: string
    NamaProgramStudi: string
    KodePendaftar: string
    NoUjian: string
    Periode: string
    MataKuliah: number
    Status: string
    SanggahanAssesmenId: string
    SanggahanProsesBanding: boolean
    SanggahanDiskusiBanding: boolean
}